requirements.txt
elasticsearch
psycopg2
pyyaml
pytz
django
django-haystack
